Refactor function_schema module
- Feature check instead of version number
- Move utility functions into `utils` module
- Move `Doc` into `types` module.
